+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.1.0.nullsafety.0]
+
+* Migrates to null-safety
+* Renames `Router -> PathRouter`
+* Refactors `PathRouteHandler.pattern`
+  Makes it a getter property initialized at constructor (final)
+
 ## [0.0.1+3]
 
 * Removes warnings for flutter stable 1.17